fixed steam build flow errors

+++ bash/steam/install_steamcmd.sh
@@ -1,18 +1,33 @@
 #!/bin/sh
 
+source bash/util/package_util.sh
+
 # installing yay for steamcmd
 echo " > installing steamcmd requirements"
-sudo rm /etc/ld.so.conf.d/fakeroot.conf
-sudo pacman -S fakeroot
+if [ -f "/etc/ld.so.conf.d/fakeroot.conf" ]; then
+    sudo rm /etc/ld.so.conf.d/fakeroot.conf
+fi
 # sudo pacman -S base-devel # this should be installed prior?
-sudo pacman -S pacman
-sudo pacman -S libarchive
+# install_library "pacman"
+install_library "fakeroot"
+install_library "libarchive"
+install_library "makepkg"
 
+# if on arch linux
 echo " > installing yay"
 cd $HOME/projects
 git clone https://aur.archlinux.org/yay.git
 cd yay
 makepkg -si
+
+# if ubuntu with apt
+sudo apt install -y git perl
+cd $HOME/projects
+git clone https://aur.archlinux.org/trizen.git
+cd trizen
+makepkg -si
+
+
 
 echo " > installing steamcmd with yay"
 yay -S steamcmd
